Add more compiler warning output and a few more things (#278)
* Add more compiler warning output. Fix new warnings. Add Werror option. Use Werror in CI. Remove hardcoded CPU core number in CI.

+++ amr-wind/utilities/FieldPlaneAveraging.cpp
@@ -108,7 +108,7 @@ FPlaneAveraging<FType>::line_average_interpolated(amrex::Real x, int comp) const
     int ind = 0;
 
     if (x > m_xlo + 0.5 * m_dx) {
-        ind = floor((x - m_xlo) / m_dx - 0.5);
+        ind = static_cast<int>(floor((x - m_xlo) / m_dx - 0.5));
         const amrex::Real x1 = m_xlo + (ind + 0.5) * m_dx;
         c = (x - x1) / m_dx;
     }
@@ -198,7 +198,7 @@ amrex::Real FPlaneAveraging<FType>::line_derivative_interpolated(
     int ind = 0;
 
     if (x > m_xlo + 0.5 * m_dx) {
-        ind = floor((x - m_xlo) / m_dx - 0.5);
+        ind = static_cast<int>(floor((x - m_xlo) / m_dx - 0.5));
         const amrex::Real x1 = m_xlo + (ind + 0.5) * m_dx;
         c = (x - x1) / m_dx;
     }
@@ -405,7 +405,7 @@ VelPlaneAveraging::line_hvelmag_average_interpolated(amrex::Real x) const
     int ind = 0;
 
     if (x > m_xlo + 0.5 * m_dx) {
-        ind = floor((x - m_xlo) / m_dx - 0.5);
+        ind = static_cast<int>(floor((x - m_xlo) / m_dx - 0.5));
         const amrex::Real x1 = m_xlo + (ind + 0.5) * m_dx;
         c = (x - x1) / m_dx;
     }
